fix: logging
